pyproject.toml
[tool.poetry]
name = "star-wars-character-classifier"
version = "0.1.0"
description = "Star Wars character classifier (Lightning + MLflow + DVC + Hydra + ONNX)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "star_wars_characters" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.4"
pillow = "^10.4.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
fire = "^0.6.0"
dvc = "^3.55.0"
onnxruntime = "^1.18.1"

[tool.poetry.group.train.dependencies]
torch = "^2.3.1"
torchvision = "^0.18.1"
lightning = "^2.3.3"
torchmetrics = "^1.4.1"
mlflow = "^2.14.3"
matplotlib = "^3.9.0"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
onnx = "^1.16.1"
kaggle = "^1.6.17"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
pre-commit = "^3.7.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
prettier = "^0.0.7"

[tool.poetry.scripts]
swc = "star_wars_characters.cli:main"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
